Paginas piden login

+++ web/application/views/productos_view.php
@@ -3,7 +3,9 @@
     <meta http-equiv='Content-Type' content='text/html; charset=utf-8'/>
     <title><?php echo $titulo;?></title>
 </head>
-<header>Productos</header>
+<h1>Productos</h1>
+<h2>Bienvenido <?=$this->session->userdata('username')?></h2>
+<p><?=anchor(base_url().'login/logout_ci', 'Cerrar sesión')?></p>
 <?php
 if(isset($actualizarProducto)){
     $ID = '<p><input type="hidden" name="ID" value="'.$this->uri->segment(3).'"></p>';
@@ -20,6 +22,7 @@ if(isset($actualizarProducto)){
 }
 ?>
 <body>
+<?php if($this->session->userdata('perfil')=='Administrador'){?>
 <form action="<?php echo base_url();?>productos/<?php echo $action;?>" method="post">
     <?php echo $ID; ?>
     <p><label>Clave:</label><input type="text" name="Clave" value="<?php echo $Clave?>"/></p>
@@ -33,14 +36,17 @@ if(isset($actualizarProducto)){
     }
     ?>
 </form>
+<?php } ?>
 <?php if(count($productosGuardadas)>0):?>
     <?php foreach($productosGuardadas as $producto) : ?>
         <table>
             <tr><?php echo $producto->ID; ?> -- </tr>
             <tr><?php echo $producto->Clave; ?>     </tr>
             <tr><?php echo $producto->Descripcion; ?>     </tr>
-            <tr><a href="<?php echo base_url(); ?>productos/index/<?php echo $producto->ID; ?>">modificar    </a></tr>
-            <tr><a href="<?php echo base_url(); ?>productos/eliminar/<?php echo $producto->ID; ?>">eliminar</a></p></tr>
+            <?php if($this->session->userdata('perfil')=='Administrador'){?>
+                <tr><a href="<?php echo base_url(); ?>productos/index/<?php echo $producto->ID; ?>">modificar    </a></tr>
+                <tr><a href="<?php echo base_url(); ?>productos/eliminar/<?php echo $producto->ID; ?>">eliminar</a></p></tr>
+            <?php } ?>
         </table>
     <?php endforeach; ?>
 <?php else :?>
