+++ web/application/views/tarjetas_view.php
@@ -3,7 +3,9 @@
     <meta http-equiv='Content-Type' content='text/html; charset=utf-8'/>
     <title><?php echo $titulo;?></title>
 </head>
-<header>Tarjetas</header>
+<h1>Tarjetas</h1>
+<h2>Bienvenido <?=$this->session->userdata('username')?></h2>
+<p><?=anchor(base_url().'login/logout_ci', 'Cerrar sesión')?></p>
 <?php
 if(isset($actualizarTarjeta)){
     $ID = '<p><input type="hidden" name="ID" value="'.$this->uri->segment(3).'"></p>';
@@ -18,6 +20,7 @@ if(isset($actualizarTarjeta)){
 }
 ?>
 <body>
+<?php if($this->session->userdata('perfil')=='Administrador'){?>
 <form action="<?php echo base_url();?>tarjetas/<?php echo $action;?>" method="post">
     <?php echo $ID; ?>
     <p><label>Descripcion:</label><input type="text" name="Descripcion" value="<?php echo $Descripcion?>"/></p>
@@ -30,13 +33,16 @@ if(isset($actualizarTarjeta)){
     }
     ?>
 </form>
+<?php } ?>
 <?php if(count($tarjetasGuardadas)>0):?>
     <?php foreach($tarjetasGuardadas as $tarjeta) : ?>
         <table>
             <tr><?php echo $tarjeta->ID; ?> -- </tr>
             <tr><?php echo $tarjeta->Descripcion; ?>     </tr>
-            <tr><a href="<?php echo base_url(); ?>tarjetas/index/<?php echo $tarjeta->ID; ?>">modificar    </a></tr>
-            <tr><a href="<?php echo base_url(); ?>tarjetas/eliminar/<?php echo $tarjeta->ID; ?>">eliminar</a></p></tr>
+            <?php if($this->session->userdata('perfil')=='Administrador'){?>
+                <tr><a href="<?php echo base_url(); ?>tarjetas/index/<?php echo $tarjeta->ID; ?>">modificar    </a></tr>
+                <tr><a href="<?php echo base_url(); ?>tarjetas/eliminar/<?php echo $tarjeta->ID; ?>">eliminar</a></p></tr>
+            <?php } ?>
         </table>
     <?php endforeach; ?>
 <?php else :?>
