+++ web/application/views/usos_view.php
@@ -3,7 +3,9 @@
     <meta http-equiv='Content-Type' content='text/html; charset=utf-8'/>
     <title><?php echo $titulo;?></title>
 </head>
-<header>Usos</header>
+<h1>Usos</h1>
+<h2>Bienvenido <?=$this->session->userdata('username')?></h2>
+<p><?=anchor(base_url().'login/logout_ci', 'Cerrar sesión')?></p>
 <?php
 if(isset($actualizarUso)){
     $ID = '<p><input type="hidden" name="ID" value="'.$this->uri->segment(3).'"></p>';
@@ -26,6 +28,7 @@ if(isset($actualizarUso)){
 }
 ?>
 <body>
+<?php if($this->session->userdata('perfil')=='Administrador' || $this->session->userdata('perfil')=='Compras'){?>
 <form action="<?php echo base_url();?>usos/<?php echo $action;?>" method="post">
     <?php echo $ID; ?>
     <select name="IDMina">
@@ -46,7 +49,6 @@ if(isset($actualizarUso)){
     </select>
     <p><label>Cantidad:</label><input type="text" name="Cantidad" value="<?php echo $Cantidad?>"/></p>
     <p><input type="hidden" name="IDUsuario" value="<?=$this->session->userdata('id_usuario')?>"></p>
-    <?=$this->session->userdata('id_usuario')?>
     <p><label>Fecha de uso:</label><input type="date" name="Fecha" value="<?php echo $Fecha?>"/></p>
     <p><input type="submit" name="guardar" value="<?php echo $button?>" /></p>
     <?php
@@ -57,6 +59,7 @@ if(isset($actualizarUso)){
     }
     ?>
 </form>
+<?php } ?>
 <?php if(count($usosGuardados)>0):?>
 
     <?php foreach($usosGuardados as $uso) : ?>
@@ -67,8 +70,10 @@ if(isset($actualizarUso)){
             <tr><?php echo $uso->Cantidad; ?>  -- </tr>
             <tr><?php foreach ($usuariosGuardados as $usuarios) :  if ($uso->IDUsuario==$usuarios->ID){ echo $usuarios->username; break;} endforeach; ?>  -- </tr>
             <tr><?php echo $uso->Fecha; ?>     </tr>
-            <tr><a href="<?php echo base_url(); ?>usos/index/<?php echo $uso->ID; ?>">modificar    </a></tr>
-            <tr><a href="<?php echo base_url(); ?>usos/eliminar/<?php echo $uso->ID; ?>">eliminar</a></p></tr>
+            <?php if($this->session->userdata('perfil')=='Administrador' || $this->session->userdata('perfil')=='Compras'){?>
+                <tr><a href="<?php echo base_url(); ?>usos/index/<?php echo $uso->ID; ?>">modificar    </a></tr>
+                <tr><a href="<?php echo base_url(); ?>usos/eliminar/<?php echo $uso->ID; ?>">eliminar</a></p></tr>
+            <?php } ?>
         </table>
     <?php endforeach; ?>
 <?php else :?>
